Implement docker, nginx, and fix analysis API fetching
The client-side server can now be containerized and serve its production
build as static assets using nginx. Requests to the analysis API will
be reverse proxied by nginx.

The client used to send fetch requests to the analysis API without a
trailing slash, which resulted in failure to resolve to the API endpoint
with nginx reverse proxy. Fetch requests will now include the slash so
nginx can resolve requests to the upstream endpoint.