Implement authorization to invoke cloud run services
The nginx proxy pass used to pass requests to the FastAPI cloud run
service without an authorization header, which resulted in 403
responses.

Now, upon container startup, the entrypoint start.sh will retrieve the
appropriate authorization token and set it in the authorization header
in proxy passes to the FastAPI server.